+++ src/variables.js
@@ -37,7 +37,7 @@ exports.economyMerchMap = economyMerchMap;
 var dlogDetails = {};
 exports.dlogDetails = dlogDetails;
 
-var userSettings = { "theme": "auto", "use_custom_theme": false, "theme_background": "#ffffff", "theme_main": "#ffffff", "theme_darken_ratio": null, "notificationRefresh": 30, "unit": "metric", "radio": "disabled", "radio_type": "tsr", "radio_volume": 100, "display_timezone": Intl.DateTimeFormat().resolvedOptions().timeZone, "data_saver": false };
+var userSettings = { "theme": "auto", "use_custom_theme": false, "theme_background": "#ffffff", "theme_main": "#ffffff", "theme_darken_ratio": null, "notificationRefresh": 30, "unit": "metric", "radio": "disabled", "radio_type": "tsr", "radio_volume": 100, "display_timezone": Intl.DateTimeFormat().resolvedOptions().timeZone, "data_saver": false, "font_size": "regular" };
 // radio: enabled / disabled / auto-play (enabled)
 // radio-type: tsr / {url}
 if (localStorage.getItem("client-settings") !== null) {